Выносит моки пункта назначения в отдельный файл
