+++ src/mock/events.js
@@ -1,3 +1,4 @@
+import {destinations} from '@/mock/destinations';
 import {getRandomArrayItem, getRandomIntegerNumber, getRandomBoolean} from '@/utils/common';
 import {generateOffers} from '@/mock/offers';
 
@@ -14,20 +15,6 @@ const eventTypes = [
   `check-in`,
   `sightseeing`,
   `restaurant`
-];
-
-const destinations = [
-  `Amsterdam`,
-  `London`,
-  `Barcelona`,
-  `Machu-Pikchu`,
-  `Rome`,
-  `Pataya`,
-  `Sidney`,
-  `Melbourne`,
-  `St.Petersburg`,
-  `Vilnius`,
-  `Budapest`
 ];
 
 const getRandomStartDate = () => {
@@ -48,23 +35,6 @@ const getRandomEndDate = (startDate) => {
   return endDate;
 };
 
-const getRandomDescription = () => {
-  const template = `Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras aliquet varius magna, non porta ligula feugiat eget. Fusce tristique felis at fermentum pharetra. Aliquam id orci ut lectus varius viverra. Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante. Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum. Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui. Sed sed nisi sed augue convallis suscipit in sed felis. Aliquam erat volutpat. Nunc fermentum tortor ac porta dapibus. In rutrum ac purus sit amet tempus.`;
-  const sentenceCount = getRandomIntegerNumber(1, 5);
-
-  return template.split(`. `).slice(0, sentenceCount).join(`. `) + `.`;
-};
-
-const getPhotosList = (count) => {
-  const photosAmount = count ? count : getRandomIntegerNumber(1, 5);
-  const photosArray = [];
-  for (let i = 0; i < photosAmount; i++) {
-    photosArray[i] = `http://picsum.photos/248/152?r=${Math.random()}`;
-  }
-
-  return photosArray;
-};
-
 const generateEvent = () => {
   const isFavorite = getRandomBoolean();
   const type = getRandomArrayItem(eventTypes);
@@ -75,19 +45,16 @@ const generateEvent = () => {
 
   const offers = generateOffers(type);
 
-  const description = getRandomDescription();
-  const photos = getPhotosList();
-
   return {
     type,
     isFavorite,
-    destination,
+    destination: destination.name,
     dateStart,
     dateEnd,
     price,
     offers,
-    description,
-    photos,
+    description: destination.description,
+    photos: destination.photos,
   };
 };
